updates test kernels
